chore(opapi): mini refactor
