+++ opapi/src/create-state.ts
@@ -2,11 +2,11 @@ import type { SchemaObject } from 'openapi3-ts/oas31'
 import { VError } from 'verror'
 import { z } from 'zod'
 import { schema } from './opapi'
-import type { PathParams } from './path-params'
-import { isAlphanumeric, isCapitalAlphabetical, uniqueBy } from './util'
-import { generateSchemaFromZod } from './jsonschema'
+import type { PathParams } from './utils/path-params'
+import { isAlphanumeric, isCapitalAlphabetical, uniqueBy } from './utils/util'
+import { generateSchemaFromZod } from './generate-schema-from-zod'
 import { OpenApiZodAny } from '@anatine/zod-openapi'
-import { objects } from './objects'
+import { objects } from './utils/objects'
 
 type SchemaType = 'zod-schema' | 'json-schema'
 type SchemaOfType<T extends SchemaType> = T extends 'zod-schema' ? OpenApiZodAny : SchemaObject
