+++ opapi/src/section-types-generator/generator.ts
@@ -1,6 +1,6 @@
 import { camel } from 'radash'
-import { saveFile } from 'src/file'
-import log from 'src/log'
+import { saveFile } from '../utils/file'
+import log from '../utils/log'
 import * as helpers from './helpers'
 import { Block, BlockComposer, DefaultState, OperationParser, SectionParser } from './types'
