+++ opapi/src/section-types-generator/parsers.ts
@@ -1,7 +1,7 @@
 import { compile as compileSchemaToTypes } from 'json-schema-to-typescript'
 import { SchemaObject } from 'openapi3-ts/oas31'
 import { title } from 'radash'
-import { Operation, isOperationWithBodyProps } from 'src/state'
+import { Operation, isOperationWithBodyProps } from '../create-state'
 import { addPropertyToBlock, getBlankBlock, pascalize, remove$RefPropertiesFromSchema } from './helpers'
 import { Block, OperationParser, SectionParser } from './types'
