+++ opapi/src/utils/path-params.ts
